@@ -4,13 +4,15 @@
   "version": "0.3.12",
   "author": "Matrix.org (original fork from Martyn Smith <martyn@dollyfish.net.nz>)",
   "scripts": {
+    "prepare": "npm run build",
+    "build": "tsc --project tsconfig.json",
     "test": "./node_modules/faucet/bin/cmd.js test/test-*.js",
     "lint": "./node_modules/jscs/bin/jscs --preset=airbnb */*.js"
   },
   "contributors": [
     "node-irc contributors"
   ],
-  "types": "index.d.ts",
+  "types": "lib/index.d.ts",
   "repository": {
     "type": "git",
     "url": "http://github.com/matrix-org/node-irc"
@@ -18,20 +20,21 @@
   "bugs": {
     "url": "http://github.com/matrix-org/node-irc/issues"
   },
-  "main": "lib/irc",
+  "main": "lib/index",
   "engines": {
-    "node": ">=0.10.0"
+    "node": ">=12.0.0"
   },
   "license": "GPL-3.0",
   "dependencies": {
-    "irc-colors": "^1.5.0",
-    "utf-8-validate": "^5.0.2",
+    "chardet": "^1.3.0",
     "iconv-lite": "^0.6.2",
-    "chardet": "^1.3.0"
+    "irc-colors": "^1.5.0",
+    "utf-8-validate": "^5.0.5"
   },
   "devDependencies": {
+    "@types/node": "^14",
     "faucet": "0.0.1",
-    "jscs": "^3.0.7",
-    "tape": "^5.2.2"
+    "tape": "^5.2.2",
+    "typescript": "^4.2.4"
   }
 }
