@@ -1,20 +1,20 @@
-var parseMessage  = require('../lib/parse_message');
-var test = require('tape');
+const { parseMessage } = require('../lib/parse_message');
+const test = require('tape');
 
-var testHelpers = require('./helpers');
+const testHelpers = require('./helpers');
 
 test('irc.parseMessage', function(t) {
-    var checks = testHelpers.getFixtures('parse-line');
+    const checks = testHelpers.getFixtures('parse-line');
 
     Object.keys(checks).forEach(function(line) {
-        var stripColors = false;
+        let stripColors = false;
         if (checks[line].hasOwnProperty('stripColors')) {
             stripColors = checks[line].stripColors;
             delete checks[line].stripColors;
         }
-        t.equal(
-            JSON.stringify(checks[line]),
-            JSON.stringify(parseMessage(line, stripColors)),
+        t.deepEqual(
+            checks[line],
+            parseMessage(line, stripColors),
             line + ' parses correctly'
         );
     });
